# replace tsu.AssociativeArray with an indexed map for 3.8x perf gain, smaller package