--- a/.gitignore
+++ b/.gitignore
@@ -1,5 +1,6 @@
 node_modules
 tmp/
+es/
 /coverage
 lib/kiwi.min.js
 lib/kiwi.d.ts