--- a/.gitignore
+++ b/.gitignore
@@ -1,6 +1,5 @@
 node_modules
 tmp/
-es/
 /coverage
 lib/kiwi.min.js
 lib/kiwi.d.ts