--- a/es/constraint.d.ts
+++ b/es/constraint.d.ts
@@ -1,0 +1,103 @@
+import { Expression } from "./expression";
+import { Variable } from "./variable";
+/**
+ * Kiwi is an efficient implementation of the Cassowary constraint solving
+ * algorithm, based on the seminal Cassowary paper.
+ * It is *not* a refactoring or port of the original C++ solver, but
+ * has been designed from the ground up to be lightweight and fast.
+ *
+ * **Example**
+ *
+ * ```javascript
+ * var kiwi = require('kiwi');
+ *
+ * // Create a solver
+ * var solver = new kiwi.Solver();
+ *
+ * // Create and add some editable variables
+ * var left = new kiwi.Variable();
+ * var width = new kiwi.Variable();
+ * solver.addEditVariable(left, kiwi.Strength.strong);
+ * solver.addEditVariable(width, kiwi.Strength.strong);
+ *
+ * // Create a variable calculated through a constraint
+ * var centerX = new kiwi.Variable();
+ * var expr = new kiwi.Expression([-1, centerX], left, [0.5, width]);
+ * solver.addConstraint(new kiwi.Constraint(expr, kiwi.Operator.Eq, kiwi.Strength.required));
+ *
+ * // Suggest some values to the solver
+ * solver.suggestValue(left, 0);
+ * solver.suggestValue(width, 500);
+ *
+ * // Lets solve the problem!
+ * solver.updateVariables();
+ * assert(centerX.value(), 250);
+ * ```
+ *
+ * ##API Documentation
+ * @module kiwi
+ */
+/**
+ * An enum defining the linear constraint operators.
+ *
+ * |Value|Operator|Description|
+ * |----|-----|-----|
+ * |`Le`|<=|Less than equal|
+ * |`Ge`|>=|Greater than equal|
+ * |`Eq`|==|Equal|
+ *
+ * @enum {Number}
+ */
+export declare enum Operator {
+    Le = 0,
+    Ge = 1,
+    Eq = 2
+}
+/**
+ * A linear constraint equation.
+ *
+ * A constraint equation is composed of an expression, an operator,
+ * and a strength. The RHS of the equation is implicitly zero.
+ *
+ * @class
+ * @param {Expression} expression The constraint expression (LHS).
+ * @param {Operator} operator The equation operator.
+ * @param {Expression} [rhs] Right hand side of the expression.
+ * @param {Number} [strength=Strength.required] The strength of the constraint.
+ */
+export declare class Constraint {
+    constructor(expression: Expression | Variable, operator: Operator, rhs?: Expression | Variable | number, strength?: number);
+    /**
+     * A static constraint comparison function.
+     * @private
+     */
+    static Compare(a: Constraint, b: Constraint): number;
+    /**
+     * Returns the unique id number of the constraint.
+     * @private
+     */
+    id(): number;
+    /**
+     * Returns the expression of the constraint.
+     *
+     * @return {Expression} expression
+     */
+    expression(): Expression;
+    /**
+     * Returns the relational operator of the constraint.
+     *
+     * @return {Operator} linear constraint operator
+     */
+    op(): Operator;
+    /**
+     * Returns the strength of the constraint.
+     *
+     * @return {Number} strength
+     */
+    strength(): number;
+    toString(): string;
+    private _expression;
+    private _operator;
+    private _strength;
+    private _id;
+}