--- a/es/kiwi.d.ts
+++ b/es/kiwi.d.ts
@@ -1,0 +1,6 @@
+export * from "./constraint";
+export * from "./expression";
+export * from "./maptype";
+export * from "./solver";
+export * from "./strength";
+export * from "./variable";