--- a/es/kiwi.js
+++ b/es/kiwi.js
@@ -1,0 +1,13 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) 2014-2018, Nucleic Development Team & H. Rutjes.
+|
+| Distributed under the terms of the Modified BSD License.
+|
+| The full license is in the file COPYING.txt, distributed with this software.
+|----------------------------------------------------------------------------*/
+export * from "./constraint";
+export * from "./expression";
+export * from "./maptype";
+export * from "./solver";
+export * from "./strength";
+export * from "./variable";