--- a/thirdparty/tsu.d.ts
+++ b/thirdparty/tsu.d.ts
@@ -286,3 +286,4 @@ declare module tsu {
 }
 declare module tsu {
 }
+export = tsu;