Note: this branches from my other PR, es6-modules.  If you decide not to use that, I can re-submit against master.

This PR replaces tsu.AssociativeArray with a basic indexed map implementation, so that lookups run in O(1) rather than O(log n) time.  The new implementation recreates the parts of the tsu.AssociativeArray interface that were used by kiwi.js, requiring no code changes to consumers.  The dependency on the internal tsu library is dropped.

This results in a 1.1x improvement to the creation benchmark and a 3.8x improvement to the update benchmark.  It also removes 370 lines from the package.

```sh
silence:kiwi.js aah$ git checkout es6-modules
Switched to branch 'es6-modules'
silence:kiwi.js aah$ node -v
v10.9.0
silence:kiwi.js aah$ npm run bench

> kiwi.js@1.0.2 bench /Users/aah/Dropbox/Tech Projects/kiwi.js
> node bench/main.js

----- Running creation benchmark...
Cassowary.js x 3,073 ops/sec ±1.07% (83 runs sampled)
kiwi x 19,942 ops/sec ±0.45% (83 runs sampled)
kiwi new API x 18,855 ops/sec ±0.71% (90 runs sampled)
Fastest is kiwi (± 6.49x faster)
----- Running solving benchmark...
Cassowary.js x 193,467 ops/sec ±0.46% (88 runs sampled)
kiwi x 140,130 ops/sec ±0.58% (90 runs sampled)
Fastest is Cassowary.js (± 1.38x faster)
silence:kiwi.js aah$ git checkout indexed-map
Switched to branch 'indexed-map'
silence:kiwi.js aah$ npm run bench

> kiwi.js@1.0.2 bench /Users/aah/Dropbox/Tech Projects/kiwi.js
> node bench/main.js

----- Running creation benchmark...
Cassowary.js x 3,101 ops/sec ±1.05% (83 runs sampled)
kiwi x 22,242 ops/sec ±0.41% (90 runs sampled)
kiwi new API x 20,867 ops/sec ±0.75% (86 runs sampled)
Fastest is kiwi (± 7.17x faster)
----- Running solving benchmark...
Cassowary.js x 194,443 ops/sec ±0.50% (86 runs sampled)
kiwi x 536,907 ops/sec ±0.33% (94 runs sampled)
Fastest is kiwi (± 2.76x faster)
```

One behavioral change which I do not think is significant but which bears mentioning is that tsu.AssociativeArray kept its backing array sorted in the order of the keys' .id() property.  The indexed map does not, so the order in which items are iterated will be different.  I don't believe this is significant, but I'm still digesting the cassowary algorithm.  All tests pass.